+++ src/js/core/overflow-auto.js
@@ -6,23 +6,23 @@ export default {
     mixins: [Class],
 
     props: {
-        selModal: String,
-        selPanel: String,
+        selContainer: String,
+        selContent: String,
     },
 
     data: {
-        selModal: '.uk-modal',
-        selPanel: '.uk-modal-dialog',
+        selContainer: '.uk-container',
+        selContent: '.uk-modal-dialog',
     },
 
     computed: {
 
-        modal({selModal}, $el) {
-            return closest($el, selModal);
+        container({selContainer}, $el) {
+            return closest($el, selContainer);
         },
 
-        panel({selPanel}, $el) {
-            return closest($el, selPanel);
+        content({selContent}, $el) {
+            return closest($el, selContent);
         }
 
     },
@@ -35,13 +35,13 @@ export default {
 
         read() {
 
-            if (!this.panel || !this.modal) {
+            if (!this.content || !this.container) {
                 return false;
             }
 
             return {
                 current: toFloat(css(this.$el, 'maxHeight')),
-                max: Math.max(150, height(this.modal) - (offset(this.panel).height - height(this.$el)))
+                max: Math.max(150, height(this.container) - (offset(this.content).height - height(this.$el)))
             };
         },